+++ app/keepers/keepers.go
@@ -649,9 +649,8 @@ func (ak *AppKeepers) InitKeepers(
 	ak.BTCStkConsumerKeeper = bsckeeper.NewKeeper(
 		appCodec,
 		runtime.NewKVStoreService(keys[bsctypes.StoreKey]),
-		ak.AccountKeeper,
-		ak.BankKeeper,
 		ak.IBCKeeper.ClientKeeper,
+		ak.IBCKeeper.ChannelKeeper,
 		ak.WasmKeeper,
 		appparams.AccGov.String(),
 	)