+++ x/btcstaking/keeper/msg_server.go
@@ -371,7 +371,7 @@ func (ms msgServer) AddCovenantSigs(goCtx context.Context, req *types.MsgAddCove
 
 	// All is fine add received signatures to the BTC delegation and BtcUndelegation
 	// and emit corresponding events
-	ms.addCovenantSigsToBTCDelegation(
+	if err := ms.addCovenantSigsToBTCDelegation(
 		ctx,
 		delInfo,
 		req.Pk,
@@ -380,7 +380,9 @@ func (ms msgServer) AddCovenantSigs(goCtx context.Context, req *types.MsgAddCove
 		parsedUnbondingSlashingAdaptorSignatures,
 		req.StakeExpansionTxSig,
 		btcTip.Height,
-	)
+	); err != nil {
+		return nil, err
+	}
 
 	// at this point, the covenant signatures are verified and are not duplicated.
 	// Thus, we can safely consider this message as refundable
@@ -648,7 +650,9 @@ func (ms msgServer) BTCUndelegate(goCtx context.Context, req *types.MsgBTCUndele
 
 	// all good, add the signature to BTC delegation's undelegation
 	// and set back
-	ms.btcUndelegate(ctx, btcDel, delegatorUnbondingInfo, req.StakeSpendingTx, req.StakeSpendingTxInclusionProof)
+	if err := ms.btcUndelegate(ctx, btcDel, delegatorUnbondingInfo, req.StakeSpendingTx, req.StakeSpendingTxInclusionProof); err != nil {
+		return nil, err
+	}
 
 	// At this point, the unbonding signature is verified.
 	// Thus, we can safely consider this message as refundable