+++ x/incentive/keeper/reward_tracker.go
@@ -4,10 +4,11 @@ import (
 	"context"
 	"errors"
 
-	"github.com/babylonlabs-io/babylon/v3/x/incentive/types"
+	sdkmath "cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	sdkmath "cosmossdk.io/math"
+	bbntypes "github.com/babylonlabs-io/babylon/v3/types"
+	"github.com/babylonlabs-io/babylon/v3/x/incentive/types"
 )
 
 // AddFinalityProviderRewardsForBtcDelegations gets the current finality provider rewards
@@ -25,8 +26,10 @@ func (k Keeper) AddFinalityProviderRewardsForBtcDelegations(ctx context.Context,
 		return types.ErrFPCurrentRewardsWithoutVotingPower.Wrapf("fp %s doesn't have positive voting power", fp.String())
 	}
 
-	fpCurrentRwd.AddRewards(rwd)
-	return k.setFinalityProviderCurrentRewards(ctx, fp, fpCurrentRwd)
+	if err := fpCurrentRwd.AddRewards(rwd); err != nil {
+		return err
+	}
+	return k.SetFinalityProviderCurrentRewards(ctx, fp, fpCurrentRwd)
 }
 
 // BtcDelegationActivated adds new amount of active satoshi to the delegation
@@ -198,10 +201,14 @@ func (k Keeper) calculateDelegationRewardsBetween(
 		panic("negative rewards should not be possible")
 	}
 
-	rewardsWithDecimals := differenceWithDecimals.MulInt(btcDelRwdTracker.TotalActiveSat)
+	rewardsWithDecimals, err := bbntypes.CoinsSafeMulInt(differenceWithDecimals, btcDelRwdTracker.TotalActiveSat)
+	if err != nil {
+		return sdk.Coins{}, err
+	}
+
 	// note: necessary to truncate so we don't allow withdrawing more rewardsWithDecimals than owed
 	// QuoInt already truncates
-	rewards := rewardsWithDecimals.QuoInt(types.DecimalAccumulatedRewards)
+	rewards := rewardsWithDecimals.QuoInt(types.DecimalRewards)
 	return rewards, nil
 }
 
@@ -234,7 +241,8 @@ func (k Keeper) IncrementFinalityProviderPeriod(ctx context.Context, fp sdk.AccA
 
 	currentRewardsPerSat := sdk.NewCoins()
 	if !fpCurrentRwd.TotalActiveSat.IsZero() {
-		currentRewardsPerSatWithDecimals := fpCurrentRwd.CurrentRewards.MulInt(types.DecimalAccumulatedRewards)
+		// the fp current rewards is already set with decimals
+		currentRewardsPerSatWithDecimals := fpCurrentRwd.CurrentRewards
 		currentRewardsPerSat = currentRewardsPerSatWithDecimals.QuoInt(fpCurrentRwd.TotalActiveSat)
 	}
 
@@ -250,7 +258,7 @@ func (k Keeper) IncrementFinalityProviderPeriod(ctx context.Context, fp sdk.AccA
 
 	// initiates a new period with empty rewards and the same amount of active sat
 	newCurrentRwd := types.NewFinalityProviderCurrentRewards(sdk.NewCoins(), fpCurrentRwd.Period+1, fpCurrentRwd.TotalActiveSat)
-	if err := k.setFinalityProviderCurrentRewards(ctx, fp, newCurrentRwd); err != nil {
+	if err := k.SetFinalityProviderCurrentRewards(ctx, fp, newCurrentRwd); err != nil {
 		return 0, err
 	}
 
@@ -269,7 +277,7 @@ func (k Keeper) initializeFinalityProvider(ctx context.Context, fp sdk.AccAddres
 
 	// set current rewards (starting at period 1)
 	newFp := types.NewFinalityProviderCurrentRewards(sdk.NewCoins(), 1, sdkmath.ZeroInt())
-	return newFp, k.setFinalityProviderCurrentRewards(ctx, fp, newFp)
+	return newFp, k.SetFinalityProviderCurrentRewards(ctx, fp, newFp)
 }
 
 // initializeBTCDelegation creates a new BTCDelegationRewardsTracker from the